requirements.txt
opencv-python==4.8.1.78
Pillow==10.1.0
qrcode==7.4.2
numpy==1.26.2
pytest==7.4.3
pytest-cov==4.1.0 